nfdiv-1133: make notification methods stateless
