+++ src/main/java/uk/gov/hmcts/divorce/citizen/notification/conditionalorder/AppliedForConditionalOrderNotification.java
@@ -7,9 +7,7 @@ import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrderQuestions;
 import uk.gov.hmcts.divorce.divorcecase.model.Gender;
-import uk.gov.hmcts.divorce.notification.ApplicantNotification;
 import uk.gov.hmcts.divorce.notification.CommonContent;
-import uk.gov.hmcts.divorce.notification.NotificationService;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -17,14 +15,11 @@ import java.util.Objects;
 
 import static uk.gov.hmcts.divorce.notification.CommonContent.NO;
 import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.CITIZEN_APPLIED_FOR_CONDITIONAL_ORDER;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLIED_FOR_CONDITIONAL_ORDER;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_PARTNER_APPLIED_FOR_CONDITIONAL_ORDER;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 
 @Component
 @Slf4j
-public class AppliedForConditionalOrderNotification implements ApplicantNotification {
+public class AppliedForConditionalOrderNotification {
 
     static final String PLUS_21_DUE_DATE = "date email received plus 21 days";
     static final String PLUS_14_DUE_DATE = "application for CO date plus 14 days";
@@ -41,57 +36,10 @@ public class AppliedForConditionalOrderNotification implements ApplicantNotifica
     static final String APPLICANT1 = "applicant 1";
     static final String APPLICANT2 = "applicant 2";
 
-    private String submittingUser;
-
-    @Autowired
-    private NotificationService notificationService;
-
     @Autowired
     private CommonContent commonContent;
 
-    @Override
-    public void sendToApplicant1(final CaseData caseData, final Long id) {
-        if (APPLICANT1.equalsIgnoreCase(submittingUser)) {
-            log.info("Notifying applicant 1 that their conditional order application has been submitted: {}", id);
-            notificationService.sendEmail(
-                caseData.getApplicant1().getEmail(),
-                caseData.getApplicationType().isSole() ? CITIZEN_APPLIED_FOR_CONDITIONAL_ORDER : JOINT_APPLIED_FOR_CONDITIONAL_ORDER,
-                templateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2(), APPLICANT1),
-                caseData.getApplicant1().getLanguagePreference()
-            );
-        } else if (!alreadyApplied(caseData, APPLICANT1)) {
-            log.info("Notifying applicant 1 that their partner has submitted a conditional order application: {}", id);
-            notificationService.sendEmail(
-                caseData.getApplicant1().getEmail(),
-                JOINT_PARTNER_APPLIED_FOR_CONDITIONAL_ORDER,
-                partnerTemplateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2(), APPLICANT2),
-                caseData.getApplicant1().getLanguagePreference()
-            );
-        }
-    }
-
-    @Override
-    public void sendToApplicant2(final CaseData data, final Long id) {
-        if (APPLICANT2.equalsIgnoreCase(submittingUser)) {
-            log.info("Notifying applicant 2 that their conditional order application has been submitted: {}", id);
-            notificationService.sendEmail(
-                data.getApplicant2().getEmail(),
-                JOINT_APPLIED_FOR_CONDITIONAL_ORDER,
-                templateVars(data, id, data.getApplicant2(), data.getApplicant1(), APPLICANT2),
-                data.getApplicant2().getLanguagePreference()
-            );
-        } else if (!data.getApplicationType().isSole() && !alreadyApplied(data, APPLICANT2)) {
-            log.info("Notifying applicant 2 that their partner has submitted a conditional order application: {}", id);
-            notificationService.sendEmail(
-                data.getApplicant2().getEmail(),
-                JOINT_PARTNER_APPLIED_FOR_CONDITIONAL_ORDER,
-                partnerTemplateVars(data, id, data.getApplicant2(), data.getApplicant1(), APPLICANT1),
-                data.getApplicant2().getLanguagePreference()
-            );
-        }
-    }
-
-    private Map<String, String> templateVars(CaseData caseData, Long id, Applicant applicant, Applicant partner, String whichApplicant) {
+    protected Map<String, String> templateVars(CaseData caseData, Long id, Applicant applicant, Applicant partner, String whichApplicant) {
         Map<String, String> templateVars = commonContent.mainTemplateVars(caseData, id, applicant, partner);
         templateVars.put(PLUS_21_DUE_DATE,
             coQuestions(caseData, whichApplicant).getSubmittedDate().plusDays(21).format(DATE_TIME_FORMATTER));
@@ -101,11 +49,15 @@ public class AppliedForConditionalOrderNotification implements ApplicantNotifica
         return templateVars;
     }
 
-    private Map<String, String> partnerTemplateVars(CaseData data, Long id, Applicant applicant, Applicant partner, String whichPartner) {
+    protected Map<String, String> partnerTemplateVars(CaseData data, Long id, Applicant applicant, Applicant partner, String whichPartner) {
         Map<String, String> templateVars = commonContent.mainTemplateVars(data, id, applicant, partner);
         templateVars.put(PLUS_14_DUE_DATE,
             coQuestions(data, whichPartner).getSubmittedDate().plusDays(14).format(DATE_TIME_FORMATTER));
         return templateVars;
+    }
+
+    protected boolean alreadyApplied(CaseData caseData, String whichApplicant) {
+        return Objects.nonNull(coQuestions(caseData, whichApplicant).getSubmittedDate());
     }
 
     private Map<String, String> jointTemplateVars(CaseData data, Applicant partner, String whichApplicant) {
@@ -161,10 +113,6 @@ public class AppliedForConditionalOrderNotification implements ApplicantNotifica
         return templateVars;
     }
 
-    private boolean alreadyApplied(CaseData caseData, String whichApplicant) {
-        return Objects.nonNull(coQuestions(caseData, whichApplicant).getSubmittedDate());
-    }
-
     private ConditionalOrderQuestions coQuestions(CaseData caseData, String whichApplicant) {
         return whichApplicant.equalsIgnoreCase(APPLICANT1)
             ? caseData.getConditionalOrder().getConditionalOrderApplicant1Questions()
@@ -173,9 +121,5 @@ public class AppliedForConditionalOrderNotification implements ApplicantNotifica
 
     private String whichPartner(String whichApplicant) {
         return whichApplicant.equalsIgnoreCase(APPLICANT1) ? APPLICANT2 : APPLICANT1;
-    }
-
-    public void setSubmittingUser(boolean isApplicant1) {
-        submittingUser = isApplicant1 ? APPLICANT1 : APPLICANT2;
     }
 }
